requirements.txt
altair==4.2.0 ; python_version >= "3.8" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.8" and python_version < "4.0"
backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
blinker==1.5 ; python_version >= "3.8" and python_version < "4.0"
cachetools==5.3.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4"
charset-normalizer==3.0.1 ; python_version >= "3.8" and python_version < "4"
click==8.1.3 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
decorator==5.1.1 ; python_version >= "3.8" and python_version < "4.0"
entrypoints==0.4 ; python_version >= "3.8" and python_version < "4.0"
gitdb==4.0.10 ; python_version >= "3.8" and python_version < "4.0"
gitpython==3.1.30 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4"
importlib-metadata==6.0.0 ; python_version >= "3.8" and python_version < "4.0"
importlib-resources==5.10.2 ; python_version >= "3.8" and python_version < "3.9"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "4.0"
jsonschema==4.17.3 ; python_version >= "3.8" and python_version < "4.0"
markdown-it-py==2.1.0 ; python_version >= "3.8" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.8" and python_version < "4.0"
numpy==1.24.1 ; python_version < "4.0" and python_version >= "3.8"
packaging==23.0 ; python_version >= "3.8" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.8" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.8" and python_version < "4.0"
pkgutil-resolve-name==1.3.10 ; python_version >= "3.8" and python_version < "3.9"
protobuf==3.20.3 ; python_version >= "3.8" and python_version < "4.0"
pyarrow==10.0.1 ; python_version >= "3.8" and python_version < "4.0"
pydeck==0.8.0 ; python_version >= "3.8" and python_version < "4.0"
pygments==2.14.0 ; python_version >= "3.8" and python_version < "4.0"
pympler==1.0.1 ; python_version >= "3.8" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
pytz-deprecation-shim==0.1.0.post0 ; python_version >= "3.8" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.8" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.8" and python_version < "4"
rich==13.2.0 ; python_version >= "3.8" and python_version < "4.0"
semver==2.13.0 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
smmap==5.0.0 ; python_version >= "3.8" and python_version < "4.0"
streamlit==1.12.0 ; python_version >= "3.8" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.8" and python_version < "4.0"
toolz==0.12.0 ; python_version >= "3.8" and python_version < "4.0"
tornado==6.2 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.8" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.8" and python_version < "4.0"
tzlocal==4.2 ; python_version >= "3.8" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.8" and python_version < "4"
validators==0.20.0 ; python_version >= "3.8" and python_version < "4.0"
watchdog==2.2.1 ; python_version >= "3.8" and python_version < "4.0" and platform_system != "Darwin"
zipp==3.11.0 ; python_version >= "3.8" and python_version < "4.0"
